api/pyproject.toml
[tool.poetry]
name = "beachhub-api"
version = "0.1.0"
description = "BeachHub FastAPI service"
authors = ["Vibe Coding Hackton Team"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlmodel = "^0.0.16"
psycopg = {version = "^3.1.18", extras = ["binary"]}
redis = "^5.0.4"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.4.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
httpx = "^0.27.0"
ruff = "^0.4.4"
black = "^24.4.0"

[build-system]
requires = ["poetry-core>=1.5.0"]
build-backend = "poetry.core.masonry.api"
